additional fixes
