+++ game/crate.go
@@ -1,0 +1,14 @@
+package game
+
+import (
+	"bnd/engine"
+)
+
+type crate struct {
+	engine.StaticObject
+	removeCrate func(object engine.StaticObject)
+}
+
+func (c *crate) remove() {
+	c.removeCrate(c)
+}
